chore: move shared packages to pkg folder
