+++ pkg/constant/constant.go
@@ -1,0 +1,4 @@
+package constant
+
+// HeaderUserUIDKey is the header key for the User UID
+const HeaderUserUIDKey = "Jwt-Sub"
