+++ pkg/shared/client.go
